changed data type to floating point of hunger and energy statuses

+++ src/java/model/general_model/BaseCharacter.java
@@ -6,14 +6,11 @@ public abstract class BaseCharacter {
   private String    name;
   private String    surname;
   private String    lovelyPhrase;
-
   private Integer   age;
-  private Integer   hunger;
-  private Integer   energy;
-
+  private Double    hunger;
+  private Double    energy;
   private Boolean   isFeedable;
   private Boolean   isPlayable;
-
   private MoodState mood;
   private TyanType  type;
 
@@ -22,8 +19,8 @@ public abstract class BaseCharacter {
   public void setSurname(String surname)            { this.surname      = surname;      }
   public void setLovelyPhrase(String lovelyPhrase)  { this.lovelyPhrase = lovelyPhrase; }
   public void setAge(Integer age)                   { this.age          = age;          }
-  public void setHunger(Integer hunger)             { this.hunger       = hunger;       }
-  public void setEnergy(Integer energy)             { this.energy       = energy;       }
+  public void setHunger(Double hunger)              { this.hunger       = hunger;       }
+  public void setEnergy(Double energy)              { this.energy       = energy;       }
   public void setFeedable(Boolean isFeedable)       { this.isFeedable   = isFeedable;   }
   public void setPlayable(Boolean isPlayable)       { this.isPlayable   = isPlayable;   }
   public void setMood(MoodState mood)               { this.mood         = mood;         }
@@ -33,8 +30,8 @@ public abstract class BaseCharacter {
   public String getSurname()                        { return this.surname;      }
   public String getLovelyPhrase()                   { return this.lovelyPhrase; }
   public Integer getAge()                           { return this.age;          }
-  public Integer getHunger()                        { return this.hunger;       }
-  public Integer getEnergy()                        { return this.energy;       }
+  public Double getHunger()                         { return this.hunger;       }
+  public Double getEnergy()                         { return this.energy;       }
   public Boolean getFeedable()                      { return this.isFeedable;   }
   public Boolean getPlayable()                      { return this.isPlayable;   }
   public MoodState getMood()                        { return this.mood;         }
